requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8
ultralytics>=8.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Dataset Download
pycocotools>=2.0.6
fiftyone>=0.21.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
